S15.f.9: extract plugin.ts stateful AGT tool registrations (4323 → 3233 LOC)
Ninth sub-slice of the multi-PR train decomposing cli/src/plugin.ts
toward the Phase 2 §4.2 cap of 3000 LOC. Lifts the 11-tool stateful
AGT cluster (spawn lifecycle, mesh send/inbox/transfer, agent
discovery, live handoff) into core/agt-tools/agt.ts.

LOC delta:
- cli/src/plugin.ts: 4323 → 3233 (−1090)
- cli/src/core/agt-tools/agt.ts: NEW (~1130 LOC)
- Cumulative S15.f: 7139 → 3233 (−3906, 93% to cap, only 233 LOC away)

Tools moved (bodies byte-identical):
- azureclaw_spawn, azureclaw_spawn_status, azureclaw_spawn_destroy,
  azureclaw_spawn_list (spawn cluster)
- azureclaw_mesh_send, azureclaw_mesh_inbox,
  azureclaw_mesh_transfer_file, azureclaw_discover (mesh cluster)
- azureclaw_handoff_status (always),
  azureclaw_handoff_request, azureclaw_handoff_confirm
  (registry_mode=global only, conditional preserved)

State coupling:
- handoffProgress promoted from module-level let to a shared holder
  object: const handoffState: { current: HandoffProgress | null }.
  plugin.ts wrapper (_runHandoffOrchestration) and the extracted
  module both read/write handoffState.current — preserves object
  identity across module boundaries.
- AgtToolsDeps threads (log, bannerAlreadyPrinted, inbox, meshClient,
  identity, sandboxName, meshSend, handoffState,
  runHandoffOrchestration, recordMeshSession). The three accessor
  callbacks (meshClient/identity/sandboxName) are late-bound — tool
  execution observes the current module-level value, not the value
  at registration time (these can rotate on re-init/reconnect).

Helpers moved:
- POD_DEAD_PHASES + probeSubAgentAlive (used only by spawn_status)
  → core/agt-tools/agt.ts.
- Inline safeJson (already lifted to core/safe-json.ts in f.8)
  removed from register().
- Duplicate interface HandoffProgress declaration removed; type now
  imported from core/agt-handoff.ts (canonical since f.7).

No tool surface changes: names, descriptions, parameter schemas,
execute return shapes byte-identical. AGT mesh wire format
unchanged (KNOCK protocol, X3DH, signing, ratchet, chunking
thresholds). Trust + audit calls unchanged. Sandbox Dockerfile
COPYs cli/src/ + cli/dist/ as whole trees → new module ships
automatically.

Verification:
- npx tsc --noEmit clean
- npm run lint: 32 warnings (was 30, +2 for new module any types)
- npm test: 454 pass / 2 skipped (unchanged)
- npm run build: clean

Audit: docs/security-audits/2026-04-29-phase2-hotspot-plugin-cli-f9.md
